+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 -->
 
 ## Master
+- Add GitHub draft support[@captainbarbosa][] - [#341](https://github.com/danger/swift/pull/341)
 
 ## 3.3.0
 
@@ -399,3 +400,4 @@ This release also includes:
 [@fortmarek]: https://github.com/danger/swift/pull/283
 [@avdlee]: https://github.com/AvdLee
 [@lucasmpaim]: https://github.com/lucasmpaim
+[@captainbarbosa]: https://github.com/captainbarbosa